requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
filelock==3.13.1
fonttools==4.47.2
fsspec==2023.12.2
gpuctypes==0.2.0
idna==3.6
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
pyobjc-core==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-libdispatch==10.1
pyobjc-framework-Metal==10.1
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
sympy==1.12
tinygrad @ git+https://github.com/tinygrad/tinygrad.git@50f2e31d26f6af504cf01ce68054444c80fe8eb7
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
